+++ CHANGELOG.rst
@@ -15,7 +15,7 @@ Coming soon
 
 These are the major improvements we are currently working on. If there is a specific bugfix or feature you would like to see, please `create an issue <https://github.com/InstituteforDiseaseModeling/covasim/issues/new/choose>`__.
 
-- Expanded tutorials (health care workers, calibration, exercises, etc.)
+- Continued updates to vaccine and variant parameters and workflows
 - Multi-region and geographical support
 - Economics and costing analysis
 
@@ -23,6 +23,28 @@ These are the major improvements we are currently working on. If there is a spec
 ~~~~~~~~~~~~~~~~~~~~~~~
 Latest versions (3.0.x)
 ~~~~~~~~~~~~~~~~~~~~~~~
+
+Version 3.0.4 (2021-05-19)
+--------------------------
+- Fixed a bug that prevented simulations from being run *without* prognoses by age.
+- Fixed an array length mismatch for single-dose vaccines.
+- The default antibody kinetics are now a 3-part curve, with a 14-day growth, 250 day exp decay and then another exponential decay with a exponentially decaying decay parameter. This is captured in the new NAb functional form, ``nab_growth_decay``. To align with this change, NAbs are now initialized at the time of infection, so that individuals build immunity over the course of infection. 
+- Some strain parameter changes based on https://www.eurosurveillance.org/content/10.2807/1560-7917.ES.2021.26.16.2100348
+- Added strain to the infection log
+- Removed the ``rel_imm_strain`` parameter; self-immunity is now always 1.0.
+- Updated vaccine and strain parameter values based on fits to empirical data.
+- Merged multisims now use the labels from each multisim, rather than the sim labels, for plotting.
+- *GitHub info*: PR `1058 <https://github.com/amath-idm/covasim/pull/1058>`__
+
+
+Version 3.0.3 (2021-05-17)
+--------------------------
+- Added a new class, ``cv.Calibration``, that can perform automatic calibration. Simplest usage is ``sim.calibrate(calib_pars)``. Note: this requires Optuna, which is not installed by default; please install separately via ``pip install optuna``. See the updated calibration tutorial for more information.
+- Added a new result, ``known_deaths``, which counts only deaths among people who have been diagnosed.
+- Updated several vaccine and variant parameters (e.g., B1.351 and B117 cross-immunity).
+- ``sim.compute_fit()`` now returns the fit by default, and creates ``sim.fit`` (previously, this was stored in ``sim.results.fit``).
+- *Regression information*: Calls to ``sim.results.fit`` should be replaced with ``sim.fit``. The ``output`` parameter for ``sim.compute_fit()`` has been removed since it now always outputs the ``Fit`` object.
+- *GitHub info*: PR `1047 <https://github.com/amath-idm/covasim/pull/1047>`__
 
 
 Version 3.0.2 (2021-04-26)